✨ :: (#1106) 기업 상세 조회에서 event 발행

+++ jobis-application/src/main/java/team/retum/jobis/domain/company/event/RecentCompanyEvent.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class RecentCompanyEvent {
-    private final Integer userId;
-    private final Integer companyId;
+    private final Long userId;
+    private final Long companyId;
 }
